requirements.txt
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT-Extended